+++ app/src/js/actions/index.js
@@ -424,6 +424,26 @@ export const bulkGranule = (payload) => ({
   }
 });
 
+export const bulkGranuleClearError = (requestId) => ({
+  type: types.BULK_GRANULE_CLEAR_ERROR,
+  requestId
+});
+
+export const bulkGranuleDelete = (payload) => ({
+  [CALL_API]: {
+    type: types.BULK_GRANULE_DELETE,
+    method: 'POST',
+    path: 'granules/bulkDelete',
+    requestId: payload.requestId,
+    body: payload.json
+  }
+});
+
+export const bulkGranuleDeleteClearError = (requestId) => ({
+  type: types.BULK_GRANULE_DELETE_CLEAR_ERROR,
+  requestId
+});
+
 export const deleteGranule = (granuleId) => ({
   [CALL_API]: {
     type: types.GRANULE_DELETE,
@@ -952,6 +972,15 @@ export const createReconciliationReport = () => ({
     type: types.NEW_RECONCILIATION,
     method: 'POST',
     path: 'reconciliationReports'
+  }
+});
+
+export const deleteReconciliationReport = (reconciliationName) => ({
+  [CALL_API]: {
+    id: reconciliationName,
+    type: types.RECONCILIATION,
+    method: 'DELETE',
+    path: `reconciliationReports/${reconciliationName}`
   }
 });