+++ app/src/js/reducers/granules.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import { set } from 'object-path';
+import { set, del } from 'object-path';
 import assignDate from './utils/assign-date';
 import removeDeleted from './utils/remove-deleted';
 import { createReducer } from '@reduxjs/toolkit';
@@ -8,7 +8,7 @@ import { getCollectionId } from '../utils/format';
 import {
   createErrorReducer,
   createInflightReducer,
-  createSuccessReducer,
+  createSuccessReducer
 } from './utils/reducer-creators';
 
 import {
@@ -33,6 +33,11 @@ import {
   BULK_GRANULE,
   BULK_GRANULE_INFLIGHT,
   BULK_GRANULE_ERROR,
+  BULK_GRANULE_CLEAR_ERROR,
+  BULK_GRANULE_DELETE,
+  BULK_GRANULE_DELETE_INFLIGHT,
+  BULK_GRANULE_DELETE_ERROR,
+  BULK_GRANULE_DELETE_CLEAR_ERROR,
   GRANULE_DELETE,
   GRANULE_DELETE_INFLIGHT,
   GRANULE_DELETE_ERROR,
@@ -64,6 +69,7 @@ export const initialState = {
 };
 
 const getConfigRequestId = ({ config: { requestId } }) => requestId;
+const getRequestId = ({ requestId }) => requestId;
 
 export default createReducer(initialState, {
   [GRANULE]: (state, action) => {
@@ -111,9 +117,20 @@ export default createReducer(initialState, {
   [GRANULE_REMOVE]: createSuccessReducer('removed'),
   [GRANULE_REMOVE_INFLIGHT]: createInflightReducer('removed'),
   [GRANULE_REMOVE_ERROR]: createErrorReducer('removed'),
+  [BULK_GRANULE_DELETE]: createSuccessReducer('bulkDelete', getConfigRequestId),
+  [BULK_GRANULE_DELETE_INFLIGHT]: createInflightReducer('bulkDelete', getConfigRequestId),
+  [BULK_GRANULE_DELETE_ERROR]: createErrorReducer('bulkDelete', getConfigRequestId),
+  [BULK_GRANULE_DELETE_CLEAR_ERROR]: (state, action) => {
+    const requestId = getRequestId(action);
+    del(state, ['bulkDelete', requestId, 'error']);
+  },
   [BULK_GRANULE]: createSuccessReducer('bulk', getConfigRequestId),
   [BULK_GRANULE_INFLIGHT]: createInflightReducer('bulk', getConfigRequestId),
   [BULK_GRANULE_ERROR]: createErrorReducer('bulk', getConfigRequestId),
+  [BULK_GRANULE_CLEAR_ERROR]: (state, action) => {
+    const requestId = getRequestId(action);
+    del(state, ['bulk', requestId, 'error']);
+  },
   [GRANULE_DELETE]: createSuccessReducer('deleted'),
   [GRANULE_DELETE_INFLIGHT]: createInflightReducer('deleted'),
   [GRANULE_DELETE_ERROR]: createErrorReducer('deleted'),