+++ app/src/js/utils/table-config/pdrs.js
@@ -14,8 +14,8 @@ export const tableColumns = [
   },
   {
     Header: 'Name',
-    accessor: row => <Link to={`pdrs/pdr/${row.pdrName}`}>{row.pdrName}</Link>,
-    id: 'name'
+    accessor: 'pdrName',
+    Cell: ({ cell: { value } }) => <Link to={`pdrs/pdr/${value}`}>{value}</Link> // eslint-disable-line react/prop-types
   },
   {
     Header: 'Status',
@@ -57,8 +57,8 @@ export const errorTableColumns = [
   },
   {
     Header: 'Name',
-    accessor: row => <Link to={`pdrs/pdr/${row.pdrName}`}>{row.pdrName}</Link>,
-    id: 'name'
+    accessor: 'name',
+    Cell: ({ cell: { value } }) => <Link to={`pdrs/pdr/${value}`}>{value}</Link> // eslint-disable-line react/prop-types
   },
   {
     Header: 'Error',