+++ app/src/js/utils/table-config/reconciliation-reports.js
@@ -3,12 +3,13 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 
 import { nullValue, dateOnly } from '../format';
+import { getReconciliationReport, deleteReconciliationReport, listReconciliationReports } from '../../actions';
 
-export const tableColumns = [
+export const tableColumns = ({ dispatch }) => ([
   {
     Header: 'Name',
-    id: 'name',
-    accessor: row => <Link to={`/reconciliation-reports/report/${row.name}`}>{row.name}</Link>
+    accessor: 'name',
+    Cell: ({ cell: { value } }) => <Link to={`/reconciliation-reports/report/${value}`}>{value}</Link> // eslint-disable-line react/prop-types
   },
   {
     Header: 'Report Type',
@@ -20,16 +21,55 @@ export const tableColumns = [
   },
   {
     Header: 'Date Generated',
-    id: 'createdAt',
-    accessor: row => dateOnly(row.createdAt)
+    accessor: 'createdAt',
+    Cell: ({ cell: { value } }) => dateOnly(value)
   },
   {
-    Header: 'Download Report'
+    Header: 'Download Report',
+    id: 'download',
+    accessor: 'name',
+    Cell: ({ cell: { value } }) => { // eslint-disable-line react/prop-types
+      return (
+        <button className='button button__row button__row--download'
+          onClick={e => handleDownloadClick(e, value, dispatch)}
+        />
+      );
+    },
+    disableSortBy: true
   },
   {
-    Header: 'Delete Report'
+    Header: 'Delete Report',
+    id: 'delete',
+    accessor: 'name',
+    Cell: ({ cell: { value } }) => ( // eslint-disable-line react/prop-types
+      <button className='button button__row button__row--delete'
+        onClick={e => handleDeleteClick(e, value, dispatch)}
+      />
+    ),
+    disableSortBy: true
   }
-];
+]);
+
+const handleDownloadClick = (e, reportName, dispatch) => {
+  e.preventDefault();
+  dispatch(getReconciliationReport(reportName)).then(response => {
+    const { data } = response;
+    const jsonHref = `data:text/json;charset=utf-8,${encodeURIComponent(JSON.stringify(data))}`;
+    const link = document.createElement('a');
+    link.setAttribute('download', `${reportName}.json`);
+    link.href = jsonHref;
+    document.body.appendChild(link);
+    link.click();
+    link.parentNode.removeChild(link);
+  });
+};
+
+const handleDeleteClick = (e, value, dispatch) => {
+  e.preventDefault();
+  dispatch(deleteReconciliationReport(value)).then(() => {
+    dispatch(listReconciliationReports());
+  });
+};
 
 export const bulkActions = function (reports) {
   return [];
@@ -46,8 +86,8 @@ export const tableColumnsS3Files = [
   },
   {
     Header: 'S3 Link',
-    accessor: row => row ? <a href={row.path} target='_blank'>Link</a> : nullValue,
-    id: 'path'
+    accessor: 'path',
+    Cell: ({ cell: { value } }) => value ? <a href={value} target='_blank'>Link</a> : nullValue // eslint-disable-line react/prop-types
   }
 ];
 
@@ -66,8 +106,8 @@ export const tableColumnsFiles = [
   },
   {
     Header: 'S3 Link',
-    accessor: row => row ? <a href={row.path} target='_blank'>Link</a> : nullValue,
-    id: 'path'
+    accessor: 'path',
+    Cell: ({ cell: { value } }) => value ? <a href={value} target='_blank'>Link</a> : nullValue // eslint-disable-line react/prop-types
   }
 ];